Try to reorganize projects into somewhat 'vertical slices' or 'feature-based' grouping. I don't know yet if it is good idea
