+++ TaskPipeline.ApiService/DAL/MongoDbContext.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using MongoDB.Driver;
-using TaskPipeline.ApiService.Models;
+using TaskPipeline.ApiService.Pipelines;
+using TaskPipeline.ApiService.Tasks;
+using TaskPipeline.ApiService.Users;
 
 namespace TaskPipeline.ApiService.DAL;
 
@@ -14,7 +16,7 @@ public class MongoDbContext : DbContext
         _database = client.GetDatabase("YourDatabaseName");
     }
 
-    public IMongoCollection<Models.ExecutableTask> Tasks => _database.GetCollection<Models.ExecutableTask>("Tasks");
+    public IMongoCollection<ExecutableTask> Tasks => _database.GetCollection<ExecutableTask>("Tasks");
     public IMongoCollection<Pipeline> Pipelines => _database.GetCollection<Pipeline>("Pipelines");
     public IMongoCollection<User> Users => _database.GetCollection<User>("Users");
 }
