+++ TaskPipeline.ApiService/Tasks/ExecutableTaskDto.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace TaskPipeline.ApiService.Models;
+namespace TaskPipeline.ApiService.Tasks;
 
 public class ExecutableTaskDto
 {
